aws/typescript: split up lab04 into 2 labs: create cluster, use cluster
