+++ labs/aws/typescript/README.md
@@ -21,11 +21,16 @@ the use of loops to create loadbalanced infrastructure in EC2.
 In this [lab](./lab-03/README.md), you will learn about how to create and deploy container based applications to Elastic
 Container Service (ECS).
 
-### Lab 4 - Deploying Containers to a Kubernetes Cluster
+### Lab 4 - Deploying a Kubernetes Cluster
 
-In this [lab](./lab-04/README.md), you will learn about how to create and deploy container based applications to a Kubernetes
+In this [lab](./lab-04/README.md), you will learn about how to create and deploy a Kubernetes Cluster in EKS.
 Cluster.
 
-### Lab 5 - Using AWS Lambda for Serverless Application Patterns
+### Lab 5 - Deploying Containers to a Kubernetes Cluster
 
-In this [lab](./lab-05/README.md), you will learn about how to create and deploy serverless applications to AWS Lambda.
+In this [lab](./lab-05/README.md), you will learn about how to create and deploy container based applications to a Kubernetes
+Cluster.
+
+### Lab 6 - Using AWS Lambda for Serverless Application Patterns
+
+In this [lab](./lab-06/README.md), you will learn about how to create and deploy serverless applications to AWS Lambda.
