+++ labs/aws/typescript/lab-04/code/step1.ts
@@ -1,0 +1,1 @@
+import * as eks from "@pulumi/eks"
