+++ labs/aws/typescript/lab-04/code/step2.ts
@@ -1,2 +1,7 @@
-import * as k8s from "@pulumi/kubernetes";
-import * as pulumi from "@pulumi/pulumi"
+import * as eks from "@pulumi/eks"
+
+// Create an EKS cluster with the default VPC, and default node group with 
+// two t2.medium node instances.
+const cluster = new eks.Cluster("eks", {
+    deployDashboard: false,
+});
