+++ labs/aws/typescript/lab-05/code/step6.ts
@@ -1,34 +1,49 @@
-import * as AWS from "aws-sdk";
-import * as aws from "@pulumi/aws";
-import * as awsx from "@pulumi/awsx";
+import * as k8s from "@pulumi/kubernetes";
 import * as pulumi from "@pulumi/pulumi";
 
-const hits = new aws.dynamodb.Table("hits", {
-    attributes: [{ name: "Site", type: "S" }],
-    hashKey: "Site",
-    billingMode: "PAY_PER_REQUEST",
-});
+let pulumiConfig = new pulumi.Config();
 
-const site = new awsx.apigateway.API("site", {
-    routes: [{
-        path: "/",
-        method: "GET",
-        eventHandler: async () => {
-            const dc = new AWS.DynamoDB.DocumentClient();
-            const result = await dc.update({
-                TableName: hits.name.get(),
-                Key: { "Site": "ACMECorp" },
-                UpdateExpression: "SET Hits = if_not_exists(Hits, :zero) + :incr",
-                ExpressionAttributeValues: { ":zero": 0, ":incr": 1 },
-                ReturnValues: "UPDATED_NEW",
-            }).promise();
-            return {
-                statusCode: 200,
-                headers: { "Content-Type": "text/html" },
-                body: "<h1>Welcome to ACMECorp!</h1>\n"+
-                    `<p>${result.Attributes!.Hits} hits.</p>\n`,
-            };
+// Existing Pulumi stack reference in the format:
+// <organization>/<project>/<stack> e.g. "myUser/myProject/dev"
+const clusterStackRef = new pulumi.StackReference(pulumiConfig.require("clusterStackRef"));
+
+// Get the kubeconfig from the cluster stack output.
+const kubeconfig = clusterStackRef.getOutput("kubeconfig").apply(JSON.stringify);
+
+// Create the k8s provider with the kubeconfig.
+const provider = new k8s.Provider("k8sProvider", {kubeconfig});
+
+const ns = new k8s.core.v1.Namespace("app-ns", {
+    metadata: { name: "joe-duffy" },
+}, {provider});
+
+const appLabels = { app: "iac-workshop" };
+const deployment = new k8s.apps.v1.Deployment("app-dep", {
+    metadata: { namespace: ns.metadata.name },
+    spec: {
+        selector: { matchLabels: appLabels },
+        replicas: 3,
+        template: {
+            metadata: { labels: appLabels },
+            spec: {
+                containers: [{
+                    name: "iac-workshop",
+                    image: "jocatalin/kubernetes-bootcamp:v2",
+                }],
+            },
         },
-    }],
-});
-export const url = site.url;
+    },
+}, {provider});
+
+const service = new k8s.core.v1.Service("app-svc", {
+    metadata: { namespace: ns.metadata.name },
+    spec: {
+        selector: appLabels,
+        ports: [{ port: 80, targetPort: 8080 }],
+        type: "LoadBalancer",
+    },
+}, {provider});
+
+const address = service.status.loadBalancer.ingress[0].hostname;
+const port = service.spec.ports[0].port;
+export const url = pulumi.interpolate`http://${address}:${port}`;
