+++ labs/aws/typescript/lab-06/code/step1.ts
@@ -1,0 +1,4 @@
+import * as AWS from "aws-sdk";
+import * as aws from "@pulumi/aws";
+import * as awsx from "@pulumi/awsx";
+import * as pulumi from "@pulumi/pulumi";
